refactor: move mesh_api.py and mesh_manager.py to mesh folder
