+++ mesh/__init__.py
@@ -1,0 +1,3 @@
+"""
+Mesh package initialization
+"""
