+++ mesh/agents/__init__.py
@@ -1,0 +1,3 @@
+"""
+Agents package initialization
+"""
