+++ mesh/agents/allora_price_prediction_agent.py
@@ -8,8 +8,7 @@ from dotenv import load_dotenv
 
 from core.llm import call_llm_async, call_llm_with_tools_async
 from decorators import monitor_execution, with_cache, with_retry
-
-from .mesh_agent import MeshAgent
+from mesh.mesh_agent import MeshAgent
 
 logger = logging.getLogger(__name__)
 load_dotenv()
