+++ mesh/agents/duckduckgo_search_agent.py
@@ -6,8 +6,7 @@ from duckduckgo_search import DDGS
 
 from core.llm import call_llm_async, call_llm_with_tools_async
 from decorators import monitor_execution, with_cache, with_retry
-
-from .mesh_agent import MeshAgent
+from mesh.mesh_agent import MeshAgent
 
 load_dotenv()
