+++ mesh/agents/twitter_profile_agent.py
@@ -9,8 +9,7 @@ from dotenv import load_dotenv
 
 from core.llm import call_llm_async
 from decorators import monitor_execution, with_cache, with_retry
-
-from .mesh_agent import MeshAgent
+from mesh.mesh_agent import MeshAgent
 
 logger = logging.getLogger(__name__)
 load_dotenv()
